[update] refactored code
