+++ Storyboarding/Extensions/GenreVC+DS&Delegate.swift
@@ -31,7 +31,6 @@ extension GenreViewController: UICollectionViewDelegate {
     }
 }
 
-
 //MARK: collection view data
 
 extension GenreViewController: UICollectionViewDataSource {
